Progress of reusing image extraction code.
